+++ src/queue.h
@@ -19,8 +19,8 @@ extern "C" {
 #endif
 
 os_status_t osi_queue_create(os_queue_t *queue, os_queue_definition_t *def);
-os_status_t osi_queue_enqueue(os_queue_t *queue, void *message, uint16_t to);
-os_status_t osi_queue_dequeue(os_queue_t *queue, void **message, uint16_t to);
+os_status_t osi_queue_enqueue(os_queue_t *queue, void *message, uint32_t to);
+os_status_t osi_queue_dequeue(os_queue_t *queue, void **message, uint32_t to);
 os_tuple_t osi_queue_enqueue_isr(os_queue_t *queue, void *message);
 
 #if defined(__cplusplus)
